Refactor MessagingProfiler and UI headers
Remove several unused UI helpers and header dependencies, and centralize timing/atomic-max logic in MessagingProfiler.

Changes:
- Remove SKSEMenuFramework include from include/MCP.h and relocate it to src/MCP.cpp. Update SetSection/AddSectionItem calls to use string types (no c_str()).
- Remove MCP::UI::ReadOnlyField and some utility constants (mod_name, plugin_version) from headers to reduce header coupling.
- Add NowSteadyNs() and UpdateMax() helpers in src/MessagingProfiler.cpp and replace repeated steady_clock and compare_exchange loops with these helpers.
- Simplify TaggedRow kind assignment and usages of timing helpers across MessagingProfiler implementation.
- Delete legacy averaging APIs and related types (GetAverageDurations, GetTotalsAvgMs, AverageRow, GetAverageRows) from MessagingProfiler header and implementation.
- Minor fix in Hooks.cpp: remove unnecessary static_cast in GetPluginVersion.

Rationale: reduce header dependencies, eliminate duplicated time/atomic update code, and clean up unused legacy APIs.

+++ src/Hooks.cpp
@@ -17,7 +17,7 @@ namespace {
 
     double GetPluginVersion(RE::TESFile* file) {
         if (!file) return -1.0;
-        return static_cast<double>(file->version);
+        return file->version;
     }
 
     template <class Fn, class... Args>
